requirements.txt
numpy
pandas
langchain
langgraph
pydantic
groq
requests
jsonschema
python-dotenv
tqdm
scikit-learn
matplotlib
seaborn
torch
deepeval